+++ tests/testthat/test-speed.R
@@ -1535,7 +1535,10 @@ test_that("speed handles MET with unequal site dimensions", {
     c(1, 2)
   )
   expect_setequal(
-    unique(round(matrixStats::rowVars(table(design_df$treatment, design_df$site)), 3)),
+    unique(round(
+      matrixStats::rowVars(table(design_df$treatment, design_df$site)),
+      3
+    )),
     c(0.3, 0.8)
   )
   expect_equal(max(table(design_df$site_row, design_df$treatment)), 1)